refactor: address PR review points (#218)
Tightens three boundaries flagged in the PR review:

1. Runtime layer is now command-agnostic. Moved FormCommandOptions
   (CLI-only type) back to @/commands/shared/optionTypes. src/runtime/
   has zero @/commands/* imports.

2. IPC protocol owns its DTOs. Introduced @/ipc/protocol/domTypes with
   the canonical FillResult/ClickResult/PressKeyResult/ScrollResult/
   SubmitResult/RawFormData interfaces. Runtime files re-export and
   implement these. src/ipc/ has zero @/runtime/ imports.

3. Error model relocated out of presentation layer. Moved CommandError
   and related utils from @/ui/errors to @/errors. Core/runtime modules
   that throw domain errors no longer import from @/ui/. (UI/messages
   imports from some daemon/connection/runtime modules remain; full
   push-strings-out-of-core is a larger refactor deferred.)
