+++ src/commands/dom/DomElementResolver.ts
@@ -26,8 +26,8 @@
  * ```
  */
 
+import { CommandError } from '@/errors/index.js';
 import { QueryCacheManager } from '@/session/QueryCacheManager.js';
-import { CommandError } from '@/ui/errors/index.js';
 import { createLogger } from '@/ui/logging/index.js';
 import { elementAtIndexNotFoundError, indexOutOfRangeError } from '@/ui/messages/errors.js';
 import { EXIT_CODES } from '@/utils/exitCodes.js';
