+++ src/commands/dom/get.ts
@@ -20,8 +20,8 @@ import {
 } from '@/commands/dom/semanticUtils.js';
 import { runCommand } from '@/commands/shared/CommandRunner.js';
 import type { DomGetCommandOptions } from '@/commands/shared/optionTypes.js';
+import { CommandError } from '@/errors/index.js';
 import { resolveA11yNode } from '@/telemetry/a11y.js';
-import { CommandError } from '@/ui/errors/index.js';
 import { formatDomGet } from '@/ui/formatters/dom.js';
 import { elementAtIndexNotFoundError, noNodesFoundError } from '@/ui/messages/errors.js';
 import { EXIT_CODES } from '@/utils/exitCodes.js';
