+++ src/commands/dom/helpers/query.ts
@@ -7,9 +7,9 @@
 
 import { CDPConnectionError } from '@/connection/errors.js';
 import type { Protocol } from '@/connection/typed-cdp.js';
+import { CommandError } from '@/errors/index.js';
 import { callCDP } from '@/ipc/client.js';
 import type { DomQueryResult, DomGetResult, DomGetOptions, DomContext } from '@/types.js';
-import { CommandError } from '@/ui/errors/index.js';
 import { createLogger } from '@/ui/logging/index.js';
 import {
   noNodesFoundError,
