+++ src/commands/network/har.ts
@@ -13,10 +13,10 @@ import { Option, type Command } from 'commander';
 import { runCommand } from '@/commands/shared/CommandRunner.js';
 import { jsonOption } from '@/commands/shared/commonOptions.js';
 import type { NetworkHarCommandOptions } from '@/commands/shared/optionTypes.js';
+import { CommandError } from '@/errors/index.js';
 import { getSessionFilePath } from '@/session/paths.js';
 import { applyFilters, parseFilterString, validateFilterString } from '@/telemetry/filterDsl.js';
 import { buildHAR } from '@/telemetry/har/builder.js';
-import { CommandError } from '@/ui/errors/index.js';
 import { operationFailedError } from '@/ui/messages/errors.js';
 import { AtomicFileWriter } from '@/utils/atomicFile.js';
 import { EXIT_CODES } from '@/utils/exitCodes.js';
