+++ src/commands/shared/CommandRunner.ts
@@ -1,6 +1,6 @@
 import { type BaseOptions } from '@/commands/shared/optionTypes.js';
+import { CommandError, isDaemonConnectionError } from '@/errors/index.js';
 import { OutputBuilder, buildSuccessResponse } from '@/ui/OutputBuilder.js';
-import { CommandError, isDaemonConnectionError } from '@/ui/errors/index.js';
 import { daemonNotRunningError, unknownError, genericError } from '@/ui/messages/errors.js';
 import { getErrorMessage } from '@/utils/errors.js';
 import { EXIT_CODES } from '@/utils/exitCodes.js';
