+++ src/errors/CommandError.ts
