+++ src/ipc/protocol/commands.ts
@@ -5,11 +5,15 @@
  * Each command has a request schema (input) and response data schema (output).
  */
 
+import type {
+  ClickResult,
+  FillResult,
+  PressKeyResult,
+  RawFormData,
+  ScrollResult,
+  SubmitResult,
+} from '@/ipc/protocol/domTypes.js';
 import type { PageState, SessionActivity } from '@/ipc/session/types.js';
-import type { PressKeyResult, ScrollResult } from '@/runtime/dom/formFillHelpers/index.js';
-import type { SubmitResult } from '@/runtime/dom/formSubmitHelpers.js';
-import type { RawFormData } from '@/runtime/dom/formTypes.js';
-import type { ClickResult, FillResult } from '@/runtime/dom/reactEventHelpers.js';
 import type { NetworkRequest } from '@/types.js';
 
 /**
